+++ kvdb/postgres/fixture.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/btcsuite/btcwallet/walletdb"
 	embeddedpostgres "github.com/fergusstrange/embedded-postgres"
+	"github.com/lightningnetwork/lnd/kvdb/sqlbase"
 )
 
 const (
@@ -32,7 +33,7 @@ const testMaxConnections = 50
 // to be done once, because NewFixture will create random new databases on every
 // call. It returns a stop closure that stops the database if called.
 func StartEmbeddedPostgres() (func() error, error) {
-	Init(testMaxConnections)
+	sqlbase.Init(testMaxConnections)
 
 	postgres := embeddedpostgres.NewDatabase(
 		embeddedpostgres.DefaultConfig().
