+++ kvdb/sqlbase/no_sql.go
@@ -1,5 +1,5 @@
 //go:build !kvdb_postgres
 
-package postgres
+package sqlbase
 
 func Init(maxConnections int) {}
