+++ kvdb/sqlbase/readwrite_cursor.go
@@ -1,6 +1,6 @@
 //go:build kvdb_postgres
 
-package postgres
+package sqlbase
 
 import (
 	"database/sql"
